chore: refactor cache
* Move the ReadFileWithSecrets to it's own function from the cache.
* Removing the cache argument from topfconfig.LoadFromFile. Since this is in pkg/, it shouldn't depend on an internal dependency. Also topf.yaml is only read once, we don't need any caching here anyway.
