+++ pkg/config/topf_config_test.go
@@ -8,8 +8,6 @@ import (
 	"path/filepath"
 	"strings"
 	"testing"
-
-	"github.com/postfinance/topf/internal/decryption"
 )
 
 func writeTestConfig(t *testing.T, dir, content string) string {
@@ -86,7 +84,7 @@ func TestLoadFromFile(t *testing.T) {
 				}
 				yaml += "nodes:\n  - host: n1\n    role: worker\n"
 
-				cfg, _, err := LoadFromFile(writeTestConfig(t, sub, yaml), decryption.NewCache())
+				cfg, _, err := LoadFromFile(writeTestConfig(t, sub, yaml))
 				if err != nil {
 					t.Fatal(err)
 				}
@@ -108,7 +106,7 @@ configDir: _config
 nodes:
   - host: n1
     role: worker
-`), decryption.NewCache())
+`))
 		if err == nil {
 			t.Fatal("expected error for deprecated configDir field")
 		}
